fix(limited): mp-draft review follow-ups
Six items from the multiplayer-draft review.

1. **Player slot identity** (blocking). Peer subscriber attached with
   `AuthResult.player_id`, which forge-server fills with a fresh UUID
   per connection. Host's seat assignments used the username (because
   that's what `RoomPlayerInfo` carries). The two never matched →
   peers stuck in the room without ever entering the draft view.
   forge-server's connection.rs stamps `StateUpdate.from_player` with
   username, so usernames are the canonical wire identity. Peer now
   calls `attachDraftPeer(username)` from the AuthResult handler.

2. **Host teardown on disconnect**. The disconnect handler already
   detached the peer subscription but never tore down the host
   session, leaking the room-message listener if the host
   disconnected mid-draft. Calls `teardownDraftHost()` alongside
   `detachDraftPeer()` now; teardown is a no-op when no draft is
   running, so it's safe to fire unconditionally.

3. **Defensive nullish fallback**. `active.seats[0].playerSlot ?? ""`
   replaced by an explicit `hostSlot` field on `ActiveHost` set at
   draft start. Removes two `?? ""` callsites that would have
   shipped empty `fromPlayer` strings if the host's seat ever
   ended up null (shouldn't happen, but the type was telling).

4. **Parallel per-seat broadcasts**. `broadcastPerSeatStates`
   serialized N engine reads + N WebSocket sends per pick. Now it
   `Promise.all`s the seat-state fetches and the sends — pod-of-8
   draft drops from 14 sequential awaits per pick to one round of
   concurrent fetches followed by one round of concurrent sends.

5. **Rounds field in the launch dialog** (was hardcoded `3`). Now
   matches the single-player Limited hub: configurable 1-6, default
   3. Surfaces via a third Input alongside picks-per-pass and seed.

6. **Reuse the rich SetPicker**. Extracted SetPicker / SetTile /
   filterSets from `Limited.tsx` into
   `src/components/limited/SetPicker.tsx`, with the constants
   (`DRAFTABLE_SET_TYPES`, `SET_TYPE_LABELS`) split into
   `setFilters.ts` so Vite react-refresh stays happy. Limited hub
   imports them back; multiplayer draft dialog drops its bare
   `<select>` for the same picker UX (search, type-filter chips,
   recents grid, prefetching indicator). Dialog also surfaces an
   explicit empty-state when Scryfall sets haven't hydrated yet.

Also `buildSeatAssignments` flipped to module-private (it had no
external callers), and the singleton-state assumption in
`draftHost.ts` is now documented inline.

`yarn typecheck` clean. Prettier formatted.
